Move lakehouse Purview test to FabricSpark, fix QN URL encoding bug
The original test tried to create lakehouse tables via the DW SQL endpoint,
which lacks CREATE TABLE permission on lakehouses. Move the test to
tests/fabricspark/ where models run via Livy (can write to lakehouses).

Key changes:
- Add purview_sync macro to FabricSpark adapter (was missing, so
  FabricSpark users couldn't call purview_sync at all)
- Fix URL encoding bug in get_entity_by_qualified_name: qualifiedNames
  containing %252F were corrupted by the parse_qs/urlencode round-trip
  in _api_request (decoded to %2F, re-encoded as %252F in URL, but
  server decodes to %2F which doesn't match stored %252F)
- Replace T-SQL syntax with Spark SQL (current_timestamp(), string type)
- Add search-based cleanup to handle stale entities from previous runs
